requirements.txt
Django==5.0.1
django-crum==0.7.9
django-debug-toolbar==4.3.0
django-filter==25.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djangorestframework-recursive==0.1.2
drf-spectacular==0.27.2
drf-writable-nested==0.7.2
RegonAPI==1.3.1